Invalid MidiHeader structure size. #40

+++ Source/Sanford.Multimedia.Midi/Device Classes/MidiHeaderBuilder.cs
@@ -73,8 +73,8 @@ namespace Sanford.Multimedia.Midi
             MidiHeader header = new MidiHeader();
 
             // Initialize the MidiHeader.
-            header.bufferLength = BufferLength;
-            header.bytesRecorded = BufferLength;
+            header.bufferLength = (uint)BufferLength;
+            header.bytesRecorded = (uint)BufferLength;
             header.data = Marshal.AllocHGlobal(BufferLength);
             header.flags = 0;
